Rename hotkeysDialog2.tsx
