+++ packages/core/src/components/hotkeys/hotkeysDialog.tsx
